+++ src/styles/_ValidBox.scss
