+++ src/styles/styles.scss
@@ -1,5 +1,5 @@
 @import 'App';
-@import 'LetsDoThis';
+@import 'ValidBox';
 
 body, html {
   height: 100%;
